task_planner_integrated/requirements.txt
flask>=2.3.0
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"